% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtlePath.R
\name{renderTurtlePathWidget}
\alias{renderTurtlePathWidget}
\title{Shiny render function for Turtle Path ROC}
\usage{
renderTurtlePathWidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a TurtlePathWidget.}

\item{env}{The environment in which to evaluate expr (default: parent.frame()).}

\item{quoted}{Is expr a quoted expression? (default FALSE)}
}
\value{
A Shiny render function for the Turtle Path ROC widget.
}
\description{
Use this function in a Shiny server to render a Turtle Path widget.
}
\seealso{
\code{\link{TurtlePathWidgetOutput}}
}
