% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turtlePath.R
\name{TurtlePathWidgetOutput}
\alias{TurtlePathWidgetOutput}
\title{Shiny output function for Turtle Path ROC}
\usage{
TurtlePathWidgetOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output variable to read from in Shiny.}

\item{width}{Width of the widget (default '100%').}

\item{height}{Height of the widget (default '600px').}
}
\value{
A Shiny output binding for the Turtle Path ROC widget.
}
\description{
Use this function in a Shiny UI to create a placeholder for the Turtle Path widget.
}
\seealso{
\code{\link{renderTurtlePathWidget}}
}
